Add sub-minute auto-listen intervals for ShazamPi (5s, 10s, 30s pauses)
- Support fractional minutes in refresh_task auto-refresh (int → float parsing)
- Add 5s, 10s, and 30s pause options to ShazamPi settings dropdown
- Enables near-continuous listening for faster song change detection

+++ src/refresh_task.py
@@ -397,9 +397,9 @@ class RefreshTask:
         auto_refresh = self.auto_refresh_plugin_settings.get("autoRefresh")
         if auto_refresh:
             try:
-                minutes = int(auto_refresh)
+                minutes = float(auto_refresh)
                 if minutes > 0:
-                    return minutes * 60
+                    return int(minutes * 60)
             except (ValueError, TypeError):
                 pass
         return None
